+++ test/chance.test.js
@@ -11,7 +11,7 @@ describe('chance generator', () => {
   });
 
   it('should generate values using chance with params', async () => {
-    const schema = number().example('chance', {
+    const schema = number().example({ generator: 'chance' }, {
       method: 'integer',
       params: { min: 1, max: 10 }
     })
@@ -22,7 +22,7 @@ describe('chance generator', () => {
   })
 
   it('should generate values using chance without params', async () => {
-    const schema = string().example('chance', { method: 'name', });
+    const schema = string().example({ generator: 'chance' }, { method: 'name', });
     const { counts, values } = await sample(1000, () => testDataFactory.generateValid(schema));
 
     expect(counts.length).to.be.above(900);
@@ -30,7 +30,7 @@ describe('chance generator', () => {
   })
 
   it('should report missing chance generators', async () => {
-    const schema = string().example('chance', { method: 'missing', });
+    const schema = string().example({ generator: 'chance' }, { method: 'missing', });
     await expect(testDataFactory.generateValid(schema))
       .to.be.rejectedWith('The installed version of Chance does not have the \'missing\' generator');
   })