+++ test/function.test.js
@@ -11,7 +11,7 @@ describe('function generator', () => {
   });
 
   it('should generate values using the specified function', async () => {
-    const schema = number().example('function', () => Math.random());
+    const schema = number().example({ generator: 'function' }, () => Math.random());
     const { counts, values } = await sample(1000, () => testDataFactory.generateValid(schema));
 
     expect(counts.length).to.be.above(900);
@@ -19,7 +19,7 @@ describe('function generator', () => {
   })
 
   it('should accept fn as an alias for function', async () => {
-    const schema = number().example('fn', () => Math.random());
+    const schema = number().example({ generator: 'fn' }, () => Math.random());
     const { counts, values } = await sample(1000, () => testDataFactory.generateValid(schema));
 
     expect(counts.length).to.be.above(900);
@@ -27,7 +27,7 @@ describe('function generator', () => {
   })
 
   it('should supply the chance instance', async () => {
-    const schema = number().example('fn', (chance) => chance.integer());
+    const schema = number().example({ generator: 'fn' }, ({ chance }) => chance.integer());
     const { counts, values } = await sample(1000, () => testDataFactory.generateValid(schema));
 
     expect(counts.length).to.be.above(900);