+++ test/rel-date.test.js
@@ -6,7 +6,7 @@ describe('relative date generator', () => {
   it('should generate relative dates', async () => {
     const now = new Date('2020-01-01T00:00:00.000Z');
     const testDataFactory = new TestDataFactory({ now }).addMethod(mixed, 'example');
-    const schema = date().example('rel-date', { days: 1 });
+    const schema = date().example({ generator: 'rel-date' }, { days: 1 });
     const value = await testDataFactory.generateValid(schema);
     expect(value.toISOString()).to.equal(new Date('2020-01-02T00:00:00.000').toISOString())
   })
@@ -14,7 +14,7 @@ describe('relative date generator', () => {
   it('should default to no change', async () => {
     const now = new Date('2020-01-01T00:00:00.000Z');
     const testDataFactory = new TestDataFactory({ now }).addMethod(mixed, 'example');
-    const schema = date().example('rel-date');
+    const schema = date().example({ generator: 'rel-date' });
     const value = await testDataFactory.generateValid(schema);
     expect(value.toISOString()).to.equal(new Date('2020-01-01T00:00:00.000').toISOString())
   })