+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.0.0]
+### Breaking Changes
+- `mixed.example()` now takes an object as its first parameter, with optional parameters `id` and `generator`. The `id` parameter is no longer used to resolve generators.
+- `meta.sessionKey` is no longer used. Instead the example id forms the sessionKey.
+
 ## [1.1.0]
 ### Added
 - TestDataSession events